rbac/requirements_ipsecems_rbac.txt
Django==1.10
Yapsy==1.11.223
argparse==1.2.1
djangorestframework==3.4.5
drfdocs==0.0.11
gunicorn==19.6.0
netifaces==0.10.5
python-consul==0.6.1
python-ldap==2.4.27
requests==2.11.1
six==1.10.0
wsgiref==0.1.2
